+++ src/test/java/dinkplugin/notifiers/CollectionNotifierTest.java
@@ -5,6 +5,7 @@ import dinkplugin.message.NotificationBody;
 import dinkplugin.message.NotificationType;
 import dinkplugin.notifiers.data.CollectionNotificationData;
 import dinkplugin.util.ItemSearcher;
+import net.runelite.api.GameState;
 import net.runelite.api.ItemID;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -18,6 +19,8 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 class CollectionNotifierTest extends MockedNotifierTest {
+
+    private static final int TOTAL_ENTRIES = 1443;
 
     @Bind
     @InjectMocks
@@ -36,6 +39,12 @@ class CollectionNotifierTest extends MockedNotifierTest {
         when(config.notifyCollectionLog()).thenReturn(true);
         when(config.collectionSendImage()).thenReturn(false);
         when(config.collectionNotifyMessage()).thenReturn("%USERNAME% has added %ITEM% to their collection");
+
+        // init client mocks
+        when(client.getVarpValue(CollectionNotifier.COMPLETED_VARP)).thenReturn(0);
+        when(client.getVarpValue(CollectionNotifier.TOTAL_VARP)).thenReturn(TOTAL_ENTRIES);
+        when(client.getGameState()).thenReturn(GameState.LOGGED_IN);
+        notifier.onTick();
     }
 
     @Test
@@ -54,7 +63,7 @@ class CollectionNotifierTest extends MockedNotifierTest {
             false,
             NotificationBody.builder()
                 .text(String.format("%s has added %s to their collection", PLAYER_NAME, item))
-                .extra(new CollectionNotificationData(item, ItemID.SEERCULL, (long) price))
+                .extra(new CollectionNotificationData(item, ItemID.SEERCULL, (long) price, 1, TOTAL_ENTRIES))
                 .type(NotificationType.COLLECTION)
                 .build()
         );